Enhanced BaseTestCase class with register and authenticate methods.
